+++ praw/models/__init__.py
@@ -7,12 +7,13 @@ from .list.redditor import RedditorList
 from .list.trophy import TrophyList
 from .listing.domain import DomainListing
 from .listing.generator import ListingGenerator
-from .listing.listing import Listing
+from .listing.listing import Listing, ModeratorListing
 from .mod_action import ModAction
 from .preferences import Preferences
 from .reddit.collections import Collection
 from .reddit.comment import Comment
 from .reddit.emoji import Emoji
+from .reddit.inline_media import InlineGif, InlineImage, InlineMedia, InlineVideo
 from .reddit.live import LiveThread, LiveUpdate
 from .reddit.message import Message, SubredditMessage
 from .reddit.modmail import ModmailAction, ModmailConversation, ModmailMessage