+++ praw/objector.py
@@ -115,6 +115,20 @@ class Objector:
             parser = self.parsers[self._reddit.config.kinds["comment"]]
         elif "collection_id" in data.keys():
             parser = self.parsers["Collection"]
+        elif {"moderators", "moderatorIds", "allUsersLoaded", "subredditId"}.issubset(
+            data
+        ):
+            data = snake_case_keys(data)
+            moderators = []
+            for mod_id in data["moderator_ids"]:
+                mod = snake_case_keys(data["moderators"][mod_id])
+                mod["mod_permissions"] = list(mod["mod_permissions"].keys())
+                moderators.append(mod)
+            data["moderators"] = moderators
+            parser = self.parsers["moderator-list"]
+        elif "username" in data.keys():
+            data["name"] = data.pop("username")
+            parser = self.parsers[self._reddit.config.kinds["redditor"]]
         else:
             if "user" in data:
                 parser = self.parsers[self._reddit.config.kinds["redditor"]]
@@ -137,6 +151,10 @@ class Objector:
             return None
         if isinstance(data, list):
             return [self.objectify(item) for item in data]
+        if "json" in data and "errors" in data["json"]:
+            errors = data["json"]["errors"]
+            if len(errors) > 0:
+                raise RedditAPIException(errors)
         if "kind" in data and (
             "shortName" in data or data["kind"] in ("menu", "moderators")
         ):
@@ -170,10 +188,6 @@ class Objector:
             return parser.parse(data["json"]["data"], self._reddit)
         if "rules" in data:
             return self.objectify(data["rules"])
-        if "json" in data and "errors" in data["json"]:
-            errors = data["json"]["errors"]
-            if len(errors) > 0:
-                raise RedditAPIException(errors)
         elif isinstance(data, dict):
             return self._objectify_dict(data)